Move initializer to `lib.rs`
